+++ src/Filters/Fxaa.cpp
@@ -10,7 +10,7 @@ Vec3 Fxaa::getPixelAt(Vec3 *img, uint64_t x, uint64_t y, uint64_t width, uint64_
 {
 	x = std::min(width - 1, std::max((uint64_t)0, x));
 	y = std::min(height - 1, std::max((uint64_t)0, y));
-	return (img[(int64_t)(x + y * width)]);
+	return (img[x + y * width]);
 }
 
 Vec3 Fxaa::process(Vec3 *img, Vec2 &pos, uint64_t width, uint64_t height)
